% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{color_impute}
\alias{color_impute}
\title{Impute missing values}
\usage{
color_impute(im, colorNA = "#ffffff")
}
\arguments{
\item{im}{An image object of package \strong{Magick}.}

\item{colorNA}{Color in hex format (e.g. "#ffffff").}
}
\value{
An image object of package \strong{Magick}.
}
\description{
Replace missing values by a default color.
}
\examples{
im0 <- magick::image_read("https://goo.gl/nRQi5n")
im <- color_impute(im0)
}
