% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{color_type}
\alias{color_type}
\title{Check the color type}
\usage{
color_type(colors)
}
\arguments{
\item{colors}{charactor vector or list of numeric vector}
}
\value{
Character string, type of color
}
\description{
Check the color type
}
