% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{colors_kmeans}
\alias{colors_kmeans}
\title{Colors from kmeans}
\usage{
colors_kmeans(im, kmeans)
}
\arguments{
\item{im}{An image object of package \strong{Magick}.}

\item{kmeans}{An object of class \emph{kcca} of package \strong{flexclust}.}
}
\value{
A matrix of colors (in hex format, e.g. "#ffffff").
}
\description{
Get projection on kmeans for colors of image.
}
\examples{
im0 <- magick::image_read("https://goo.gl/nRQi5n")
im <- magick::image_resize(im0, "100")
kmeans <- kmeans_colors(im, 5)
im_small <- magick::image_resize(im0, "30")
plot(im_small)
table(colors <- colors_kmeans(im_small, kmeans))
plot_color_matrix(colors)
}
