% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gray.R
\name{convert_gray}
\alias{convert_gray}
\title{Convert image into gray scale}
\usage{
convert_gray(img)
}
\arguments{
\item{img}{An \code{magick-image} object.}
}
\value{
An \code{magick-image} object.
}
\description{
Convert image into gray scale
}
