% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_palette.R
\name{convert_palette}
\alias{convert_palette}
\title{Convert the color of image to the color in the palette}
\usage{
convert_palette(img, palette = pixelit_16)
}
\arguments{
\item{img}{An \code{magick-image} object.}

\item{palette}{A palette.}
}
\value{
An \code{magick-image} object.
}
\description{
Convert the color of image to the color in the palette
}
