% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsize.R
\name{downsize}
\alias{downsize}
\title{Downsize an image}
\usage{
downsize(im, npixel_width = 200)
}
\arguments{
\item{im}{An image object of package magick.}

\item{npixel_width}{Number of pixels for the width of the image.}
}
\value{
An image object of package \strong{Magick} of reduced or equal size.
}
\description{
Downsize an image
}
\examples{
im0 <- magick::image_read("https://goo.gl/nRQi5n")
print(im <- downsize(im0, 100))
print(downsize(im, 200))  # only downsize
}
