% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-colors.R
\name{hex2rgb}
\alias{hex2rgb}
\title{Hex to RGB colors}
\usage{
hex2rgb(color)
}
\arguments{
\item{color}{Color in hex format (e.g. "#ffffff").}
}
\value{
A color in RGB format (e.g. \code{c(255, 255, 255)}).
}
\description{
Hex to RGB colors
}
\examples{
hex2rgb("#ffffff")
hex2rgb("#000000")
}
