% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{kmeans_colors}
\alias{kmeans_colors}
\title{Perform K-means on RGB colors}
\usage{
kmeans_colors(im, ncolors = 8)
}
\arguments{
\item{im}{An image object of package magick.}

\item{ncolors}{Number of colors (clusters). Default is 8.}
}
\value{
An object of class \emph{kcca} of package \strong{flexclust}.
}
\description{
Perform K-means on RGB colors
}
\examples{
im0 <- magick::image_read("https://goo.gl/nRQi5n")
im <- magick::image_resize(im0, "100")
kmeans <- kmeans_colors(im)
kmeans@centers
}
