% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline}
\alias{pipeline}
\title{Pipeline to get pixel art}
\usage{
pipeline(url, resize1, ncolors, resize2, colorNA)
}
\arguments{
\item{url}{URL of image.}

\item{resize1}{Initial resize width for processing.}

\item{ncolors}{Number of different colors.}

\item{resize2}{Final resize widthg.}

\item{colorNA}{Color replacing missing values (e.g. "#ffffff").}
}
\value{
A \emph{ggplot} object.
}
\description{
Pipeline to get pixel art
}
\examples{
pipeline(
  url = "https://goo.gl/nRQi5n", 
  resize1 = 100,
  resize2 = 22,
  ncolors = 6,
  colorNA = "#ffffff"
)
}
