% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelart-package.R
\docType{package}
\name{pixelart-package}
\alias{pixelart}
\alias{pixelart-package}
\title{pixelart: Resize Images and Make Pixel Art Models From Pictures}
\arguments{
\item{im}{An image object of package magick.}
}
\description{
Resize and crop images, and make Pixel Art models from pictures (Shiny app).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://privefl.github.io/pixelart}
  \item Report bugs at \url{https://github.com/privefl/pixelart/issues}
}

}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

}
\keyword{internal}
