% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pixelit_16}
\alias{pixelit_16}
\title{Palette of 16 colours from Pixel It}
\format{
An object of class \code{list} of length 16.
}
\source{
\url{https://github.com/giventofly/pixelit}
}
\usage{
pixelit_16
}
\description{
A list of 16 numeric vectors containing RGB values for the default colour palette on Pixel It.
}
\keyword{datasets}
