% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-colors.R
\name{plot_color_matrix}
\alias{plot_color_matrix}
\title{Plot color matrix}
\usage{
plot_color_matrix(raster)
}
\arguments{
\item{raster}{A matrix of colors (in hex format, e.g. "#ffffff").}
}
\value{
A \emph{ggplot} object.
}
\description{
Plot color matrix
}
\examples{
(colors <- matrix(c("#008744", "#0057e7", "#d62d20", "#ffa700"), 2))
plot_color_matrix(colors)
}
