% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline}
\alias{pipeline}
\title{Pipeline to get pixel art}
\usage{
pipeline(input)
}
\arguments{
\item{input}{A list with
\itemize{
\item \code{url}: the url of the image,
\item \code{resize1}: the width of the initial downsize (for processing),
\item \code{resize2}: the width of the final image,
\item \code{ncolors}: the number of different colors
}}
}
\value{
A \emph{ggplot} object.
}
\description{
Pipeline to get pixel art
}
\examples{
pipeline(list(
  url = "https://goo.gl/nRQi5n", 
  resize1 = 100,
  resize2 = 22,
  ncolors = 7
))
}
